+++ .bumpversion.cfg
@@ -1,0 +1,6 @@
+[bumpversion]
+current_version = 1.1.0
+commit = True
+tag = True
+
+[bumpversion:file:bandlab_master/__init__.py]
