+++ .gitignore
@@ -1,0 +1,9 @@
+.DS_Store
+*.pyc
+*.pyo
+*.egg-info/
+dist/
+build/
+.vscode/
+.eggs
+__pycache__
