+++ bandlab_master/__init__.py
@@ -1,0 +1,3 @@
+from .__main__ import main
+
+__version__ = "1.1"
